We only need to include identical image data once
It turns out that the advertisement images that are delivered
with the Guardian's API are very often byte-for-byte identical.
So, it makes sense to only store these once.  This produces a
great saving in the size of the final book - e.g. on today
(a Saturday) it means 3.3M rather than 7.7M.